Move git credentials check into git config service
Having a separate auth service for the purpose of checking if git is configured
correctly feels like too narrow of a scope.
