+++ apps/desktop/src/lib/ai/service.ts
@@ -18,7 +18,7 @@ import { OpenAIClient } from '$lib/ai/openAIClient';
 import { buildFailureFromAny, isFailure, ok, type Result } from '$lib/ai/result';
 import { splitMessage } from '$lib/utils/commitMessage';
 import { get } from 'svelte/store';
-import type { GitConfigService } from '$lib/backend/gitConfigService';
+import type { GitConfigService } from '$lib/config/gitConfigService';
 import type { SecretsService } from '$lib/secrets/secretsService';
 import type { TokenMemoryService } from '$lib/stores/tokenMemoryService';
 import type { HttpClient } from '@gitbutler/shared/network/httpClient';
